pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = "REST-API app"
authors = [
    {name = "Kateryna",email = "tkachenko.city@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "bcrypt (<4.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.black]
line-length = 79
target-version = ['py312']
include = '\.pyi?$'


